Make the Pie Chart components even more generic
This patch make the Pie Charts components more generic so that the chat
library is only used in one place. This ensure even better consistentcy
in the visual, as well as in the logic to group the data.
