+++ components/scrapd-age-graph.js
@@ -1,7 +1,6 @@
 import Proptypes from 'prop-types';
-import styled from '@emotion/styled';
 import { connect } from 'react-redux';
-import { RadialChart } from 'react-vis';
+import ScrapdGenericPieChart from './scrapd-generic-pie-chart';
 
 function isBetween(x, min, max) {
   return x >= min && x <= max;
@@ -45,41 +44,14 @@ const ScrapdAgeGraph = props => {
     }
   }
 
-  let data = new Array();
-  for (var key in datamap) {
-    if (datamap[key] === 0) {
-      continue;
-    }
-    let tmp = new Object();
-    tmp.angle = datamap[key];
-    tmp.label = key;
-    tmp.subLabel = Number((datamap[key] * 100) / props.fatalities.length).toFixed(0) + '%';
-    data.push(tmp);
-  }
-
-  const Graph = styled.div({
-    display: 'inline-block',
-    borderRadius: '5px',
-    border: '1px solid',
-    textAlign: 'center',
-    margin: '2em 0'
-  });
-
-  const GraphTitle = styled.p({
-    fontSize: '2em',
-    textAlign: 'center'
-  });
-
   return (
-    <Graph>
-      <RadialChart data={data} width={300} height={300} showLabels={true} labelsRadiusMultiplier={1} />
-      <GraphTitle>Age distribution</GraphTitle>
-    </Graph>
+    <ScrapdGenericPieChart groupedItems={datamap} itemCount={props.fatalities.length} attribute={props.attribute} />
   );
 };
 
 ScrapdAgeGraph.propTypes = {
-  fatalities: Proptypes.array
+  fatalities: Proptypes.array,
+  attribute: Proptypes.string
 };
 
 const mapStateToProps = state => {
