+++ components/scrapd-pie-graph.js
@@ -1,7 +1,6 @@
 import Proptypes from 'prop-types';
-import styled from '@emotion/styled';
 import { connect } from 'react-redux';
-import { RadialChart } from 'react-vis';
+import ScrapdGenericPieChart from './scrapd-generic-pie-chart';
 
 const ScrapdPieGraph = props => {
   let datamap = new Object();
@@ -15,40 +14,8 @@ const ScrapdPieGraph = props => {
     }
   }
 
-  let data = new Array();
-  for (var key in datamap) {
-    let tmp = new Object();
-    tmp.angle = datamap[key];
-    tmp.label = key;
-    tmp.subLabel = Number((datamap[key] * 100) / props.fatalities.length).toFixed(0) + '%';
-    data.push(tmp);
-  }
-
-  const Graph = styled.div({
-    display: 'inline-block',
-    borderRadius: '5px',
-    border: '1px solid',
-    textAlign: 'center',
-    margin: '2em 0'
-  });
-
-  const GraphTitle = styled.p({
-    fontSize: '2em',
-    textAlign: 'center'
-  });
-
   return (
-    <Graph>
-      <RadialChart
-        data={data}
-        width={300}
-        height={300}
-        showLabels={true}
-        labelsStyle={{ backgroundColor: 'gray', color: 'yellow' }}
-        labelsRadiusMultiplier={1}
-      />
-      <GraphTitle>{props.attribute} distribution</GraphTitle>
-    </Graph>
+    <ScrapdGenericPieChart groupedItems={datamap} itemCount={props.fatalities.length} attribute={props.attribute} />
   );
 };
